signed arithmetic fix part II

+++ src/instruction/ALU/SBCI.h
@@ -28,7 +28,7 @@ namespace emulator{
             //Z = R ==0
             //C = (!RD7&&RR7) || (RR7&&R7) || (R7 && !RD7)
 
-            uint8_t R = at.memory.GP(RdVal)-KVal;
+            int8_t R = (int8_t)at.memory.GP(RdVal)-KVal;
 
             bool
                     RD3 =utils::isSet<3>(at.memory.GP(RdVal)),
